Refactors: changed Error class to utils and file_to_string to utils
